feat: fastAPI 서버
- node랑 fastAPI 서버 동시에 돌리기
